Fix bug when only first ifetch of instruction straddling page faults
If an access causes a page fault the fetched instruction bits are
random, meaning there's a chance they correspond to the first
half of a 32-bit instruction. This then gets latched in the instruction
realigner, which waits to receive more data.
The subsequent access will usually cause the same trap, but if we
increment the PC to the next page, the fault gets overwritten in the
frontend and the realigner ignorantly attaches the nonsense bits that
were fetched under the trap. The instruction struct injected into the
instruction queue then has no record of a trap.
We fix this by muxing in zeros when the ifetch indicates a fault.
Zeros are "compressed", meaning we will be able to immediately issue
any fetched faulting instruction before we forget about the fault.
There is likely a more elegant solution and this may hurt timing,
however this change is also helpful for debugging by removing
dependencies on non-deterministic bits in faulting cases.