Fix SPI waits
These were a fixed number of nops, which caused a problem when the
code sped up from running it in bootrom.
Using the TX empty status is both more correct and faster in the
common case.
Combined with caching the bootrom, this gives a ~5x speedup for
copying images from the SD card