Improve timing for PCC change check
The old implementation checked the fully calculated target capability
to see if the metadata differed from PCC. We can instead use the base
register, since this must have the same metadata as the eventual
target. We move the computation of whether we are changing the PCC
metadata into the branch unit, which has the information of the base
capability earlier.